requirements.txt
pandas
numpy
mysql-connector-python
python-dotenv
pymysql
scikit-learn